- renamed Common to common - moved ErrorContact.vue to common folder
