+++ coops-ui/src/components/AnnualReport/ResumeErrorDialog.vue
@@ -28,7 +28,7 @@
 <script lang="ts">
 import { Component, Vue, Prop, Emit } from 'vue-property-decorator'
 import { mapGetters } from 'vuex'
-import ErrorContact from '@/components/ErrorContact.vue'
+import { ErrorContact } from '@/components/common'
 
 @Component({
   computed: {
