+++ coops-ui/src/views/StandaloneOfficeAddressFiling.vue
@@ -174,7 +174,7 @@ import ResumeErrorDialog from '@/components/AnnualReport/ResumeErrorDialog.vue'
 import SaveErrorDialog from '@/components/AnnualReport/SaveErrorDialog.vue'
 
 // Components
-import { OfficeAddresses } from '@/components/Common'
+import { OfficeAddresses } from '@/components/common'
 import Certify from '@/components/AnnualReport/Certify.vue'
 import StaffPayment from '@/components/AnnualReport/StaffPayment.vue'
 import SbcFeeSummary from 'sbc-common-components/src/components/SbcFeeSummary.vue'
