+++ coops-ui/tests/unit/AnnualReport.spec.ts
@@ -11,7 +11,7 @@ import axios from '@/axios-auth'
 import store from '@/store/store'
 import AnnualReport from '@/views/AnnualReport.vue'
 import AGMDate from '@/components/AnnualReport/AGMDate.vue'
-import { OfficeAddresses } from '@/components/Common'
+import { OfficeAddresses } from '@/components/common'
 import Directors from '@/components/AnnualReport/Directors.vue'
 import Certify from '@/components/AnnualReport/Certify.vue'
 import StaffPayment from '@/components/AnnualReport/StaffPayment.vue'
