+++ coops-ui/tests/unit/StandaloneOfficeAddressFiling.spec.ts
@@ -8,7 +8,7 @@ import flushPromises from 'flush-promises'
 import axios from '@/axios-auth'
 import store from '@/store/store'
 import StandaloneOfficeAddressFiling from '@/views/StandaloneOfficeAddressFiling.vue'
-import { OfficeAddresses } from '@/components/Common'
+import { OfficeAddresses } from '@/components/common'
 import Certify from '@/components/AnnualReport/Certify.vue'
 import StaffPayment from '@/components/AnnualReport/StaffPayment.vue'
 import VueRouter from 'vue-router'
